Migrate all data over to the people table
This commit updates all the tables to migrate over Caddies, Housing and
Golfers. It does not verify that any of the front end works still,
although the volunteer front end should still work. It also comments out
most of the validation, etc. that is broken by the migration. Those will
all be fixed in another branch that is merged after this one is deployed
and migrated.