Update models and all index pages for Person
This commit updates all the models and index pages to use the new person
object. Right now the code is referring to the volunteer path, but
ultimate that should be the person path instead. Perhaps we can make
that shift when we add Golfer, Housing, Volunteer or Caddie options to
the Person edit page.